steps 1 and 2
Step 1: get a fasta file of the proteins to be used. the .py file takes one argument, a file of the list of IDs with one on each line
Step 2: use the generated fasta file to run AlphaFold. Takes two inputs, one being the list of IDs (step 1 integrated into this) and the output directory. Note: I am still unsure about the directory that contains the data required for alphafold to run. Will update soon